REFACTOR: Make head concrete & inherited by all feature sources (#237)
